requirements.txt
langchain==0.0.259
openai==0.27.8
python-dotenv==1.0.0
jinja2==3.1.2
